feat(whkd): continue execution on hotkey conflicts
This commit ensures that execution will continue if a key binding
conflict is encountered.

If a conflict is encountered, the key combination and the bound command
will be logged out, informing the user of a conflict and bubbling up the
underlying error, before continuing to try to enable the remaining key
bindings in the whkdrc file.

re #8